moved test to avoid parallel test conflicts
