+++ test/Serilog.Settings.Configuration.Tests/ConfigurationSettingsTests.cs
@@ -78,6 +78,32 @@ namespace Serilog.Settings.Configuration.Tests
         }
 
         [Fact]
+        public void ConfigurationAssembliesFromDllScanning()
+        {
+            var json = @"{
+                ""Serilog"": {            
+                    ""Using"": [""TestDummies""],
+                    ""WriteTo"": [""DummyConsole""]
+                }
+            }";
+
+            var builder = new ConfigurationBuilder().AddJsonString(json);
+            var config = builder.Build();
+            var log = new LoggerConfiguration()
+                .ReadFrom.Configuration(
+                    configuration: config,
+                    dependencyContext: null,
+                    onNullDependencyContext: ConfigurationAssemblySource.AlwaysScanDllFiles)
+                .CreateLogger();
+
+            DummyConsoleSink.Emitted.Clear();
+
+            log.Write(Some.InformationEvent());
+
+            Assert.Equal(1, DummyConsoleSink.Emitted.Count);
+        }
+
+        [Fact]
         public void SinksAreConfigured()
         {
             var json = @"{
