+++ test/Serilog.Settings.Configuration.Tests/LoggerConfigurationExtensionsTests.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using Microsoft.Extensions.Configuration;
-using Serilog.Settings.Configuration.Tests.Support;
-using TestDummies.Console;
 using Xunit;
 
 namespace Serilog.Settings.Configuration.Tests
@@ -15,32 +13,6 @@ namespace Serilog.Settings.Configuration.Tests
 
             // should not throw
             act();
-        }
-
-        [Fact]
-        public void ConfigurationAssembliesFromDllScanning()
-        {
-            var json = @"{
-                ""Serilog"": {            
-                    ""Using"": [""TestDummies""],
-                    ""WriteTo"": [""DummyConsole""]
-                }
-            }";
-
-            var builder = new ConfigurationBuilder().AddJsonString(json);
-            var config = builder.Build();
-            var log = new LoggerConfiguration()
-                .ReadFrom.Configuration(
-                    configuration: config,
-                    dependencyContext: null,
-                    onNullDependencyContext: ConfigurationAssemblySource.AlwaysScanDllFiles)
-                .CreateLogger();
-
-            DummyConsoleSink.Emitted.Clear();
-
-            log.Write(Some.InformationEvent());
-
-            Assert.Equal(1, DummyConsoleSink.Emitted.Count);
         }
     }
 }
